Refactor: 지속되는 오류로 폴더명 Gnb에서 Navbar로 수정
